preprocess_requirements.txt
###### processor requirements ######
Cython==0.23.4
numpy==1.10.4
argparse==1.2.1
h5py==2.5.0
six==1.10.0
wsgiref==0.1.2

###### printing ######
MySQL-python==1.2.5
Pillow==3.4.2
pyserial==3.2.1